Revert "Sieu Update"
This reverts commit 9fd9471012af6f2c68acbfd147b36c9868b969dc.

+++ QuizThink/src/java/Controller/ExpertAddQuestionServlet.java
@@ -35,7 +35,7 @@ public class ExpertAddQuestionServlet extends HttpServlet {
         String duration = request.getParameter("duration");
         int subjectID = Integer.parseInt(request.getParameter("subjectID"));
         int expertID = Integer.parseInt(request.getParameter("expertID"));
-        float requirement = Float.parseFloat(request.getParameter("requirement"));
+        int requirement = Integer.parseInt(request.getParameter("requirement"));
         String status = "Successfully";
 
         //DAO
