+++ QuizThink/src/java/Controller/ExpertUpdateQuestionServlet.java
@@ -32,7 +32,7 @@ public class ExpertUpdateQuestionServlet extends HttpServlet {
         String duration = request.getParameter("duration");
         int subjectID = Integer.parseInt(request.getParameter("subjectID"));
         int expertID = Integer.parseInt(request.getParameter("expertID"));
-        float requirement = Float.parseFloat(request.getParameter("requirement"));
+        int requirement = Integer.parseInt(request.getParameter("requirement"));
         String status = "Successfully";
 
         //DAO
